Reformat type model comment, implementation of payment w/o test

+++ cashierApp/src/main/java/alimentation/cashierApp/services/PaymentService.java
@@ -3,24 +3,29 @@ package alimentation.cashierApp.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import alimentation.cashierApp.dao.PaymentRepository;
+import alimentation.cashierApp.exception.CashierAppException;
 import alimentation.cashierApp.models.Payment;
+import alimentation.cashierApp.models.Transaction;
 
 @Service
 public class PaymentService {
 
     @Autowired
     private PaymentRepository PaymentRepository;
+    @Autowired
+    private TransactionService transactionService;
 
 
     // Add methods for Payment-related operations here
-    Iterable<Payment> getAllPayments(){
+    public Iterable<Payment> getAllPayments(){
         return PaymentRepository.findAll();
     }
 
-    Payment getPaymentById(int id){
+    public Payment getPaymentById(int id){
         Optional<Payment> response = PaymentRepository.findById(id);
         if (response.isPresent()) {
             return response.get();
@@ -28,15 +33,25 @@ public class PaymentService {
         return null;
     }
 
-    void addPayment(Payment Payment){
-        PaymentRepository.save(Payment);
+    public Iterable<Payment> getAllPaymentsByTransactionId(int transactionId){
+        Transaction transaction = transactionService.getTransactionById(transactionId);
+        return PaymentRepository.findAllByTransaction(transaction);
     }
 
-    void updatePayment(Payment Payment){
-        PaymentRepository.save(Payment);
+    public Payment addPayment(Payment Payment){
+        return PaymentRepository.save(Payment);
     }
 
-    void deletePayment(int id){
+    public Payment updatePayment(Payment Payment){
+        return PaymentRepository.save(Payment);
+    }
+
+    public Payment deletePayment(int id){
+        Payment target = PaymentRepository.findById(id).get();
+        if (target == null){
+            throw new CashierAppException(HttpStatus.BAD_REQUEST, "Payment missing");
+        }
         PaymentRepository.deleteById(id);
+        return target;
     }
 }
